Organized backend routes an limited the number of times we create a new prisma client so all routes use the same client
