+++ backend/prisma/prisma.ts
@@ -1,0 +1,6 @@
+import { PrismaClient } from "@prisma/client";
+import { withAccelerate } from "@prisma/extension-accelerate";
+
+export const prisma = new PrismaClient({
+  omit: { user: { password: true } },
+}).$extends(withAccelerate());
