+++ backend/src/routes/prompt.ts
@@ -1,8 +1,6 @@
 import express from "express";
-import { PrismaClient } from "@prisma/client";
+import { prisma } from "../../prisma/prisma";
 import { getResponseForPrompt } from "./gemini";
-
-const prisma = new PrismaClient();
 
 export const promptRouter = express.Router();
